# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  position: string;
  color: string;
}