# lib/api-zod/src/generated/types/employeeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';
import type { MonthHours } from './monthHours';
import type { ObjectHours } from './objectHours';

export interface EmployeeStats {
  employee: Employee;
  totalDays: number;
  totalHours: number;
  vacationDays: number;
  byObject: ObjectHours[];
  byMonth: MonthHours[];
}