# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryInputType } from './entryInputType';
import type { Segment } from './segment';

export interface EntryInput {
  employeeId: number;
  date: string;
  type: EntryInputType;
  segments?: Segment[];
}