# lib/api-zod/src/generated/types/entryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryUpdateType } from './entryUpdateType';
import type { Segment } from './segment';

export interface EntryUpdate {
  type?: EntryUpdateType;
  segments?: Segment[];
}