# lib/api-zod/src/generated/types/recentEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentEntry {
  date: string;
  employeeName: string;
  objectName: string;
  startTime: string;
  endTime: string;
  hours: number;
}