# lib/api-zod/src/generated/types/reportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportRow } from './reportRow';

export interface ReportResult {
  rows: ReportRow[];
  totalHours: number;
  totalDays: number;
}