# lib/api-zod/src/generated/types/segment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Segment {
  objectId: number;
  startTime: string;
  endTime: string;
  /** @nullable */
  note?: string | null;
}