# lib/api-zod/src/generated/types/workObjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tabель — Work Time Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkObjectUpdate {
  /** @minLength 1 */
  name?: string;
  code?: string;
  status?: string;
}